feat: Add tensor dimension inference with comprehensive documentation and testing
Implements compile-time dimension tracking for tensors with range constraints,
enabling shape inference through operations with runtime verification and memory
estimation.

Features:
- Dimension types: Literal, Var, Named, Dynamic, Broadcast
- Shape inference: matmul, broadcast, reshape, transpose, reductions
- Runtime verification of dimension constraints
- Memory estimation from dimension bounds (min/max)
- Lean 4 formal verification of correctness

Implementation:
- Core dimension inference model (450 LOC)
- TypedTensor API with DimSpec (350 LOC)
- Lean proof generator (200 LOC)

Documentation:
- User guide with examples and troubleshooting (580 LOC)
- Design documentation with architecture details (650 LOC)
- Feature specification with BDD tests (360 LOC)
- Completion report with full deliverables (2,100 LOC)

Testing:
- Unit tests: 50+ test cases covering all dimension types
- Integration tests: 10+ real-world workflows (MLP, CNN, Transformer)
- Examples: 10 runnable examples from basic to advanced

Verification:
- Lean 4 project structure with lakefile and toolchain
- TensorDimensions.lean with 8 theorems
- TensorMemory.lean with memory safety proofs

+++ src/db/src/drivers/libsql.rs
@@ -131,7 +131,7 @@ impl LibsqlConnection {
             // Get column names
             let columns: Vec<String> = (0..column_count)
                 .map(|i| {
-                    rows.column_name(i as i32)
+                    rows.column_name(i)
                         .map(|s| s.to_string())
                         .unwrap_or_default()
                 })
@@ -144,7 +144,7 @@ impl LibsqlConnection {
                     Ok(Some(row)) => {
                         let values: Vec<SqlValue> = (0..column_count)
                             .map(|i| {
-                                let value = row.get_value(i as i32).unwrap_or(LibsqlValue::Null);
+                                let value = row.get_value(i).unwrap_or(LibsqlValue::Null);
                                 from_libsql_value(value)
                             })
                             .collect();
